Plot multiple images

Plot multiple images

Usage

plot_multi_images(list_images, par_ROWS, par_COLS, ...)

Arguments

list_images

a list of images that should be visualized

par_ROWS

an integer specifying the number of rows of the ouput plot-grid

par_COLS

an integer specifying the number of columns of the output plot-grid

...

further arguments for the 'plot_multi_images' method of the 'GaborFeatureExtract' R6 class ('OpenImageR' package)

Value

it doesn't return an R object but it displays a list of input images

Details

For the usage of the 'plot_multi_images()' function see the example section of the 'fastGLCM_Rcpp()' function and 'fastglcm()' R6 class